refactor: split cards.py into focused card module package
Reorganize the 1,930-line shitty_ui/components/cards.py into 8 focused
modules under shitty_ui/components/cards/ package:

- __init__.py: strip_urls(), get_sentiment_style() + re-exports
- empty_states.py: error cards, empty chart/HTML states
- hero.py: hero signal card
- metrics.py: metric card, performance summary
- signal.py: signal card, unified signal card
- feed.py: feed signal card, new signals banner, expandable thesis
- post.py: post card
- timeline.py: prediction timeline card, related asset link

All 18 exported symbols are re-exported from __init__.py so existing
imports (from components.cards import X) work unchanged. 104 card tests
and 152 layout tests pass with zero modifications to consumer files.
